Code review fixes: extract EventArgs, remove stray file
- Move HostNameChangedEventArgs to own file per codebase convention
- Make HostNameChangedEventArgs sealed per convention (ActiveHostChangedEventArgs)
- Remove accidentally committed gh_user.txt
- Change Dispose test to async Task with await instead of .Wait()
